Task 5 (0x0C. C - More malloc, free)

+++ 0x0C-more_malloc_free/101-mul.c
@@ -11,7 +11,7 @@ int _isdigit(char *s);
  */
 int main(int argc, char *argv[])
 {
-	long int num1, num2, mul;
+	unsigned int num1, num2, mul;
 	int i;
 
 	if (argc != 3)
@@ -32,7 +32,7 @@ int main(int argc, char *argv[])
 	num1 = atoi(argv[1]);
 	num2 = atoi(argv[2]);
 	mul = num1 * num2;
-	printf("%ld\n", mul);
+	printf("%u\n", mul);
 
 	return (0);
 }
